dom/imptests/webapps/DOMCore/tests/approved/test_Range-deleteContents.html

<!doctype html>
<title>Range.deleteContents() tests</title>
<link rel=author title="Aryeh Gregor" href=ayg@aryeh.name>
<p>To debug test failures, look at the URL of the hidden iframes corresponding
to the test at the end of the document, load those individually, and use your
browser's debugging tools.
<div id=log></div>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=common.js></script>
<script>
"use strict";

testDiv.parentNode.removeChild(testDiv);

function myDeleteContents(range) {
	// "If the context object's start and end are the same, abort this method."
	if (range.startContainer == range.endContainer
	&& range.startOffset == range.endOffset) {
		return;
	}

	// "Let original start node, original start offset, original end node, and
	// original end offset be the context object's start and end nodes and
	// offsets, respectively."
	var originalStartNode = range.startContainer;
	var originalStartOffset = range.startOffset;
	var originalEndNode = range.endContainer;
	var originalEndOffset = range.endOffset;

	// "If original start node and original end node are the same, and they are
	// a Text or Comment node, call deleteData(original start offset, original
	// end offset − original start offset) on that node, and abort these
	// steps."
	if (originalStartNode == originalEndNode
	&& (range.startContainer.nodeType == Node.TEXT_NODE
	|| range.startContainer.nodeType == Node.COMMENT_NODE)) {
		originalStartNode.deleteData(originalStartOffset, originalEndOffset - originalStartOffset);
		return;
	}

	// "Let nodes to remove be a list of all the Nodes that are contained in
	// the context object, in tree order, omitting any Node whose parent is
	// also contained in the context object."
	//
	// We rely on the fact that the contained nodes must lie in tree order
	// between the start node, and the end node's last descendant (inclusive).
	var nodesToRemove = [];
	var stop = nextNodeDescendants(range.endContainer);
	for (var node = range.startContainer; node != stop; node = nextNode(node)) {
		if (isContained(node, range)
		&& !(node.parentNode && isContained(node.parentNode, range))) {
			nodesToRemove.push(node);
		}
	}

	// "If original start node is an ancestor container of original end node,
	// set new node to original start node and new offset to original start
	// offset."
	var newNode;
	var newOffset;
	if (originalStartNode == originalEndNode
	|| originalEndNode.compareDocumentPosition(originalStartNode) & Node.DOCUMENT_POSITION_CONTAINS) {
		newNode = originalStartNode;
		newOffset = originalStartOffset;
	// "Otherwise:"
	} else {
		// "Let reference node equal original start node."
		var referenceNode = originalStartNode;

		// "While reference node's parent is not null and is not an ancestor
		// container of original end node, set reference node to its parent."
		while (referenceNode.parentNode
		&& referenceNode.parentNode != originalEndNode
		&& !(originalEndNode.compareDocumentPosition(referenceNode.parentNode) & Node.DOCUMENT_POSITION_CONTAINS)) {
			referenceNode = referenceNode.parentNode;
		}

		// "Set new node to the parent of reference node, and new offset to one
		// plus the index of reference node."
		newNode = referenceNode.parentNode;
		newOffset = 1 + indexOf(referenceNode);
	}

	// "If original start node is a Text or Comment node, run deleteData() on
	// it, with original start offset as the first argument and (length of
	// original start node − original start offset) as the second."
	if (originalStartNode.nodeType == Node.TEXT_NODE
	|| originalStartNode.nodeType == Node.COMMENT_NODE) {
		originalStartNode.deleteData(originalStartOffset, nodeLength(originalStartNode) - originalStartOffset);
	}

	// "For each node in nodes to remove, in order, remove node from its
	// parent."
	for (var i = 0; i < nodesToRemove.length; i++) {
		nodesToRemove[i].parentNode.removeChild(nodesToRemove[i]);
	}

	// "If original end node is a Text or Comment node, run deleteData(0,
	// original end offset) on it."
	if (originalEndNode.nodeType == Node.TEXT_NODE
	|| originalEndNode.nodeType == Node.COMMENT_NODE) {
		originalEndNode.deleteData(0, originalEndOffset);
	}

	// "Set the context object's start and end to (new node, new offset)."
	range.setStart(newNode, newOffset);
	range.setEnd(newNode, newOffset);
}

function testDeleteContents(i) {
	// We only load the expected once the actual is finished, to avoid race
	// conditions.  First set the onload to null and the src to the empty
	// string, so that when we set the src for real we're not just changing the
	// hash -- otherwise the load event won't fire.  We never do anything after
	// a src set except in an onload handler, to avoid race conditions.
	var actual = actualIframes[i];
	actual.id = i + "-actual";
	actual.onload = function() { onLoadActual(i); };
	actual.src = "Range-test-iframe.html#" + testRanges[i];
}

function onLoadActual(i) {
	var expected = expectedIframes[i];
	expected.id = i + "-expected";
	expected.onload = function() { onLoadExpected(i); };
	expected.src = "Range-test-iframe.html#" + testRanges[i];
}

function onLoadExpected(i) {
	var actualRange = actualIframes[i].contentWindow.testRange;
	var expectedRange = expectedIframes[i].contentWindow.testRange;
	var actualRoots, expectedRoots;

	domTests[i].step(function() {
		assert_equals(actualIframes[i].contentWindow.unexpectedException, null,
			"Unexpected exception thrown when setting up Range for actual deleteContents()");
		assert_equals(expectedIframes[i].contentWindow.unexpectedException, null,
			"Unexpected exception thrown when setting up Range for simulated deleteContents()");
		assert_equals(typeof actualRange, "object",
			"typeof Range produced in actual iframe");
		assert_equals(typeof expectedRange, "object",
			"typeof Range produced in expected iframe");

		// Just to be pedantic, we'll test not only that the tree we're
		// modifying is the same in expected vs. actual, but also that all the
		// nodes originally in it were the same.  Typically some nodes will
		// become detached when the algorithm is run, but they still exist and
		// references can still be kept to them, so they should also remain the
		// same.
		//
		// We initialize the list to all nodes, and later on remove all the
		// ones which still have parents, since the parents will presumably be
		// tested for isEqualNode() and checking the children would be
		// redundant.
		var actualAllNodes = [];
		var node = furthestAncestor(actualRange.startContainer);
		do {
			actualAllNodes.push(node);
		} while (node = nextNode(node));

		var expectedAllNodes = [];
		var node = furthestAncestor(expectedRange.startContainer);
		do {
			expectedAllNodes.push(node);
		} while (node = nextNode(node));

		actualRange.deleteContents();
		myDeleteContents(expectedRange);

		actualRoots = [];
		for (var j = 0; j < actualAllNodes.length; j++) {
			if (!actualAllNodes[j].parentNode) {
				actualRoots.push(actualAllNodes[j]);
			}
		}

		expectedRoots = [];
		for (var j = 0; j < expectedAllNodes.length; j++) {
			if (!expectedAllNodes[j].parentNode) {
				expectedRoots.push(expectedAllNodes[j]);
			}
		}

		for (var j = 0; j < actualRoots.length; j++) {
			if (!actualRoots[j].isEqualNode(expectedRoots[j])) {
				var msg = j ? "detached node #" + j : "tree root";
				msg = "Actual and expected mismatch for " + msg + ".  ";

				// Find the specific error
				var actual = actualRoots[j];
				var expected = expectedRoots[j];

				while (actual && expected) {
					assert_equals(actual.nodeType, expected.nodeType,
						msg + "First difference: differing nodeType");
					assert_equals(actual.nodeName, expected.nodeName,
						msg + "First difference: differing nodeName");
					assert_equals(actual.nodeValue, expected.nodeValue,
						msg + 'First difference: differing nodeValue (nodeName = "' + actual.nodeName + '")');
					assert_equals(actual.childNodes.length, expected.childNodes.length,
						msg + 'First difference: differing number of children (nodeName = "' + actual.nodeName + '")');
					actual = nextNode(actual);
					expected = nextNode(expected);
				}

				assert_unreached("DOMs were not equal but we couldn't figure out why");
			}

			if (j == 0) {
				// Clearly something is wrong if the node lists are different
				// lengths.  We want to report this only after we've already
				// checked the main tree for equality, though, so it doesn't
				// mask more interesting errors.
				assert_equals(actualRoots.length, expectedRoots.length,
					"Actual and expected DOMs were broken up into a different number of pieces by deleteContents() (this probably means you created or detached nodes when you weren't supposed to)");
			}
		}
	});
	domTests[i].done();

	positionTests[i].step(function() {
		assert_equals(actualIframes[i].contentWindow.unexpectedException, null,
			"Unexpected exception thrown when setting up Range for actual deleteContents()");
		assert_equals(expectedIframes[i].contentWindow.unexpectedException, null,
			"Unexpected exception thrown when setting up Range for simulated deleteContents()");
		assert_equals(typeof actualRange, "object",
			"typeof Range produced in actual iframe");
		assert_equals(typeof expectedRange, "object",
			"typeof Range produced in expected iframe");
		assert_true(actualRoots[0].isEqualNode(expectedRoots[0]),
			"The resulting DOMs were not equal, so comparing positions makes no sense");

		assert_equals(actualRange.startContainer, actualRange.endContainer,
			"startContainer and endContainer must always be the same after deleteContents()");
		assert_equals(actualRange.startOffset, actualRange.endOffset,
			"startOffset and endOffset must always be the same after deleteContents()");
		assert_equals(expectedRange.startContainer, expectedRange.endContainer,
			"Test bug!  Expected startContainer and endContainer must always be the same after deleteContents()");
		assert_equals(expectedRange.startOffset, expectedRange.endOffset,
			"Test bug!  Expected startOffset and endOffset must always be the same after deleteContents()");

		assert_equals(actualRange.startOffset, expectedRange.startOffset,
			"Unexpected startOffset after deleteContents()");
		// How do we decide that the two nodes are equal, since they're in
		// different trees?  Since the DOMs are the same, it's enough to check
		// that the index in the parent is the same all the way up the tree.
		// But we can first cheat by just checking they're actually equal.
		assert_true(actualRange.startContainer.isEqualNode(expectedRange.startContainer),
			"Unexpected startContainer after deleteContents(), expected " +
			expectedRange.startContainer.nodeName.toLowerCase() + " but got " +
			actualRange.startContainer.nodeName.toLowerCase());
		var currentActual = actualRange.startContainer;
		var currentExpected = expectedRange.startContainer;
		var actual = "";
		var expected = "";
		while (currentActual && currentExpected) {
			actual = indexOf(currentActual) + "-" + actual;
			expected = indexOf(currentExpected) + "-" + expected;

			currentActual = currentActual.parentNode;
			currentExpected = currentExpected.parentNode;
		}
		actual = actual.substr(0, actual.length - 1);
		expected = expected.substr(0, expected.length - 1);
		assert_equals(actual, expected,
			"startContainer superficially looks right but is actually the wrong node if you trace back its index in all its ancestors (I'm surprised this actually happened");
	});
	positionTests[i].done();
}

setup({timeout: 60000});

// First test a detached Range, synchronously
test(function() {
	var range = document.createRange();
	range.detach();
	assert_throws("INVALID_STATE_ERR", function() {
		range.deleteContents();
	}, "Detached Range must throw INVALID_STATE_ERR on .deleteContents()");
}, "Detached Range");

var domTests = [];
var positionTests = [];
var actualIframes = [];
var expectedIframes = [];
for (var i = 0; i < testRanges.length; i++) {
	domTests.push(async_test("Resulting DOM for range " + i + " " + testRanges[i]));
	positionTests.push(async_test("Resulting cursor position for range " + i + " " + testRanges[i]));

	actualIframes.push(document.createElement("iframe"));
	actualIframes[i].style.display = "none";
	document.body.appendChild(actualIframes[i]);

	expectedIframes.push(document.createElement("iframe"));
	expectedIframes[i].style.display = "none";
	document.body.appendChild(expectedIframes[i]);

	testDeleteContents(i);
}
</script>
